Layout fixed, wefact factuur toegevoegd

+++ OffertTemplateTool/DAL/Models/EstimateLines.cs
@@ -14,10 +14,10 @@ namespace OffertTemplateTool.DAL.Models
         [Required]
         public string Specification { get; set; }
         [Required]
-        public decimal HourCost { get; set; }
+        public float HourCost { get; set; }
         [Required]
         public double Hours { get; set; }
         [Required]
-        public  decimal TotalCost { get; set; }
+        public  float TotalCost { get; set; }
     }
 }
